requirements.txt
# ========================================
# Automation Mail Dependencies
# ========================================
# Install with: pip install -r requirements.txt

# CLI Framework
click>=8.0.0

# Beautiful Console Output
rich>=13.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Template Rendering
Jinja2>=3.0.0

# Scheduling (for background job processing)
schedule>=1.2.0

# ========================================
# Development Dependencies
# ========================================
# Uncomment these if you're contributing

# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
